refactor: split reconciliation plan/mod.rs into focused submodules
Separates plan building, file discovery, hashing, and managed file
handling into dedicated modules (builder.rs, user_files.rs, hashing.rs,
managed_files.rs), removing the catch-all helpers.rs.

Closes #396
