+++ .centy/issues/b715f3e0-780d-4809-a2a2-295ffb20a8ff.md
@@ -1,10 +1,10 @@
 ---
 # This file is managed by Centy. Use the Centy CLI to modify it.
 displayNumber: 396
-status: open
+status: closed
 priority: 2
 createdAt: 2026-04-04T11:55:33.497169+00:00
-updatedAt: 2026-04-04T11:55:33.497169+00:00
+updatedAt: 2026-04-13T18:50:00.381065+00:00
 ---
 
 # Split reconciliation plan/mod.rs — separate plan building, file discovery, and hashing
